Upload Tax Receipt API Works
The server can handle POST requests for uploading the PDF Tax Receipt file. It means: the server can already run the python script used for decoding the PDF Tax Receipt and can return either error messages or the rfc & date from owner's tax receipt depending if the request is successful or if the tax receipt already exists or if there was an error either in the python script or in the database.