Search Person Section Works Perfectly
All the logic for the Search Person section has been completed. Maybe I should add the possibility of searching a person by typing two names like: 'Dani Sol' because for now you only can search by inserting one word either it could be the rfc or firstname or secondname or surname or secondname but you can not search for a firstname and secondname at the same time. You can not search by inserting 2 words. But the rest is fine. That's the only thing missing