Search By More than One Word Works
Searching for a name by inserting more than 1 word already works. It doesn't matter if the user types the surname first and then the first_name. It doesn't matter the order of the words. (I.E it will return the same person if it searchs either Daniel Solorio or Solorio Daniel). The only thing missing is validating the search input. Because it is not allowed to enter empty spaces or strings that have more than 5 words