Seach Input Validation Completed
In the search input you can't type blank spaces at the beginning of the string, you can't add double blank spaces between words. Also, the search request to the server is not made if the only thing that was iserted by the user was a blank space. There is a drawback, when there is no match in the database and the user keeps typing, requests to the server are still made. A good approach could be preventing sending more requests if we already know there is no match with the first string the user typed