Delete Tax Receipt Function
The Tax Receipt can be deleted after inserting the reason why it should be deleted. The functionality in the Database already works, it creates a new register in the 'deleted_tax_receipts' table and deletes the register of the selected Tax Receipt in the 'tax_receipt' table. The notification is missing and fixing the Bug when there is no Tax Receipt remaining in the table, it must show the message 'No Matches', but instead it shows the headers of the table