Deleted Tax Receipts Layout Bug Fixed
The layout in the deleted tax recepits section has been fixed, now it has a min-width, in order to not shrink the table and its content. Also, this min-width has been applied to other sections like Departments and Specific Info About a Person. Now, I have to add the delete functionality to the Delete Button in the Tax Receipt Section when you select a Specific Person